Pan Card Tampering/Pan Card Tampering Flask App/requirements.txt
Flask==1.1.1
Keras==2.4.0
opencv-python==4.4.0.46
Pillow==8.0.1
tensorflow==2.3.1
scikit-image==0.17.2